Kludgy support for Perl 5.8, using dark magick from namespace::clean.
What we're essentially doing here is waiting until the end of our lexical
scope, and then blasting away the subroutine we installed.  This
is currently broken in many ways, but works in the basic cases.

We really shouldn't just be blasting it; instead we should be restoring
the previous sub that was there (if any).