In determining whether tests based on truncate() are worth running, truncate.t calls truncate() inside an eval, which should handle the case where the function doesn't exist.  But since the return value is not checked, it doesn't handle the case where a simple truncate operation fails.  The attached patch remedies that.
This came up because on VMS, the C library's ftruncate() does not play
nice with temporary files having the delete-on-close bit set, which is
what File:Temp gives us by default.  An alternative would be to have
it roll its own temp file, but that seems an unnecessary burden to
place on a test script that is doing the right thing.