Updated File::Copy hints to check for false only.
At the time of writing, File::Copy would always return 0
to signal failure on Unix systems, but copy and cp (only)
would return "" to signal failur on Windows system.

We now don't care what sort of wacky false value File::Copy decides
to return, as long as it does so.