Checking for filename sensitive leakguards being cached.
When we install a leakguard, it includes the file that 'use autodie' was called in.
If that leakguard gets cached, then installing the cached copy in a different file
will result in autodie no longer working there. The leak guard spots a different
filename, and immediately reverts back to using the CORE behaviour.