Fatal: Smarter handling of CORE subs in nested autodie
Previously, when nesting autodie with itself, autodie would wrap its
own wrapper.  With this patch, autodie properly detects it is dealing
with a "wrapped CORE sub" and simply generates a normal CORE wrapper.

Besides removing one level of indirection, it also removes the
overhead of checking for hints for the CORE sub (which was somewhat
counterintuitive, since autodie normally ignores hints for CORE subs).

Signed-off-by: Niels Thykier <niels@thykier.net>