Fatal: Reduce the overhead of import and _make_fatal
Refactor import and _make_fatal so they "use less 'CPU';".  In
particular, many calls to prototype("CORE::sub") has been replaced
with a cache, which turns out to be cheaper in the long run.

Combined these changes appear to reduce the runtime of benchmark.pl by
about 1/8 (~4.0s to ~3.5s).

Signed-off-by: Niels Thykier <niels@thykier.net>