Fatal: Deprecate direct usage of :lexical
Currently autodie is implemented as use Fatal qw(:lexical).  Users
have previously been allowed to invoke Fatal directly with this tag
and it would behave like autodie (including throwing
autodie::exception rather than "Fatal" messages).

To soften the split of autodie and Fatal (see issue #60), we now
deprecate the use of Fatal with :lexical.  The autodie pragma is still
allowed to do it (and so is it subclasses via autodie).

Closes: #61

Signed-off-by: Niels Thykier <niels@thykier.net>