t/import-into.t: enforce the minimum version of Import::Into
Since release 2.27 we specify the mimimum version of Import::Into in META.{yml,json} as test/recommends.
See https://rt.cpan.org/Ticket/Display.html?id=101377
But as this is just a `recommends` dependency, the CPAN client is free to ignore it and not upgrade `Import::Into`. If `Import::Into` is too old, we want to skip the test in the same way as if `Import::Into` is not here at all. So we now check the $VERSION.