Remove Test::Perl::Critic as a testing requirement.
We don't appear to need the line in dist.ini specifying Test::Perl::Critic
as we already have an author test called t/critic.t. While it is nice that
this test is created automajically by Dist::Zilla::Plugin::Test::Perl::Critic,
this would make the test not run during automated testing which is desirable.

Note that this is now a build_requires test and I think that will be enough
to break the circular dep.

Resolves #89